+++ arro3-core/python/arro3/core/__init__.py
@@ -1,32 +1,4 @@
-from ._core import (
-    Array,
-    ArrayReader,
-    ChunkedArray,
-    DataType,
-    Field,
-    RecordBatch,
-    RecordBatchReader,
-    Schema,
-    Table,
-    fixed_size_list_array,
-    list_array,
-    struct_array,
-    ___version,  # noqa,
-)
+from ._core import *
+from ._core import ___version
 
 __version__: str = ___version()
-
-__all__ = (
-    "Array",
-    "ArrayReader",
-    "ChunkedArray",
-    "DataType",
-    "Field",
-    "RecordBatch",
-    "RecordBatchReader",
-    "Schema",
-    "Table",
-    "fixed_size_list_array",
-    "list_array",
-    "struct_array",
-)
